% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert gene list to igraph object}
\usage{
as_igraph(genes, minscore = 0)
}
\arguments{
\item{genes}{character vector of valid ensembl gene IDs (no version suffix)}

\item{minscore}{numeric between 0 and 1, minimal score for gene-gene
interactions to be used}
}
\value{
igraph object
}
\description{
Convert a character vector of ensembl gene IDs to an igraph object using the
pre-stored gene-gene interactions as undirected edges.
Interactions can be filtered by their evidence score.
}
\examples{
as_igraph(c("ENSG00000130203", "ENSG00000189058"),  0.9)

}
