% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_external_names}
\alias{get_external_names}
\title{Convert ENSEMBL IDs to external gene names}
\usage{
get_external_names(genes)
}
\arguments{
\item{genes}{character vector of valid ENEMBL gene IDs (no version suffix)}
}
\value{
character vector of matching external gene names
}
\description{
Convert ENSEMBL IDs to external gene names
}
\examples{
get_external_names("ENSG00000278801")

}
