% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Pruning of a gene list by interaction distance to seed genes}
\usage{
prune(genes, seed_genes, minscore = 0, maxedgedistance = Inf)
}
\arguments{
\item{genes}{character vector of ensembl IDs}

\item{seed_genes}{character vector of ensembl IDs}

\item{minscore}{minimal evidence score for edges to include (see \code{\link[=as_igraph]{as_igraph()}})}

\item{maxedgedistance}{only genes that are connected to any of the seed genes
via \code{maxedgedistance} or less edges (interactions) are retained}
}
\value{
a character vector with retained genes, always containes all seed genes
}
\description{
Prune a list of genes by maximal edge distance from provided seed genes using
gene-gene functional interactions as edges.
}
\examples{
prune(c("ENSG00000130203", "ENSG00000189058"), c("ENSG00000130203"), 0.9, 2)

}
