% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_reactome_pathways.R
\name{query_participating_genes}
\alias{query_participating_genes}
\title{Query reactome for all participating genes in a set of pathways}
\usage{
query_participating_genes(pathways)
}
\arguments{
\item{pathways}{character vector of valid reactome IDs}
}
\value{
character vector of ensembl IDs of the genes participating in any of the
input pathways.
}
\description{
Get all genes participating in the provided list of reactome pathway identifiers.
}
\examples{
query_participating_genes("R-HSA-425407")

}
