% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_reactome_pathways.R
\name{query_reactome_pathways}
\alias{query_reactome_pathways}
\title{Query reactome for involved biological pathways}
\usage{
query_reactome_pathways(genes)
}
\arguments{
\item{genes}{character vector of valid ensembl gene IDs (no version suffix)}
}
\value{
character vector of reactome IDs of the pathways involving at least
one of the input genes.
}
\description{
Get all reactome.org pathways that involve any of the genes provided.
}
\examples{
query_reactome_pathways(c("ENSG00000130203", "ENSG00000189058"))

}
