% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_shiny_app.R
\name{start_shiny_app}
\alias{start_shiny_app}
\title{Interactive pathway curation}
\usage{
start_shiny_app()
}
\description{
Start the interactive shiny app fo pathway cluster curation.
}
\examples{
\dontrun{start_shiny_app()}

}
