% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tbl_ensembl}
\alias{tbl_ensembl}
\title{ENSEMBL ID to gene name}
\format{\describe{
\item{external_gene_name}{gene name}
\item{ensembl_gene_id}{ENSEMBL gene ID}
}}
\source{
\url{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}
}
\usage{
tbl_ensembl
}
\description{
This dataset is used to map ENSEMBL IDs to gene names.
It is parsed from the biomaRt bioconductor package using ensebl version 97.
}
\keyword{datasets}
