% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tbl_ensembl2reactome}
\alias{tbl_ensembl2reactome}
\title{ENSEMBL ID to reactome.org pathway}
\format{\describe{
\item{ensembl_id}{ENSEMBL gene ID}
\item{reactome_id}{reactome.org ID for gene}
\item{reactome_pathway_id}{reactome.org ID for pathway}
\item{description}{reactome.org description of pathway}
}}
\source{
\url{https://reactome.org/download/current/Ensembl2Reactome.txt}
}
\usage{
tbl_ensembl2reactome
}
\description{
This dataset is used to map ENSEMBL IDs to reactome.org pathway IDs.
It is parsed from a downloaded file of an identfier mapping from
https://reactome.org/download.
}
\keyword{datasets}
