% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tbl_reactome_interactions}
\alias{tbl_reactome_interactions}
\title{Function gene-gene interactions}
\format{\describe{
\item{score}{evidence score}
\item{a}{ENSEMBL gene ID of interactor a}
\item{b}{ENSEMBL gene ID of interactor b}
}}
\source{
\url{http://cpws.reactome.org/caBigR3WebApp2018/FIsInGene_122718_with_annotations.txt.zip}
}
\usage{
tbl_reactome_interactions
}
\description{
This dataset contains the predicted functional gene-gene interactions.
The methodology is described in
}
\details{
Wu, G., Feng, X., & Stein, L. (2010). A human functional protein interaction
network and its application to cancer data analysis. Genome biology, 11(5), R53.
}
\keyword{datasets}
