% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_pathway}
\title{Plot a gene network graph}
\usage{
plot_graph(
  igraph,
  seed_genes,
  title = NULL,
  titlesize = 10,
  textsize = 3,
  pointsize = 1,
  edgesize = 0.5,
  layout = "stress",
  layout_args = list()
)

plot_pathway(genes, seed_genes, minscore = 0, titlesize = 10, minnodes = 1)
}
\arguments{
\item{igraph}{an igraph object as returned by \code{\link[=as_igraph]{as_igraph()}}}

\item{seed_genes}{character vector of ENSEMBL IDs of the seed genes in the igraph object}

\item{title}{plot title}

\item{titlesize}{text size for the title}

\item{textsize}{text size for labels}

\item{pointsize}{point size}

\item{edgesize}{edge strength}

\item{layout}{ggraph layout}

\item{layout_args}{list of argments to pass to ggraph layout function}

\item{genes}{character vector of ensembl IDs of the genes in the network}

\item{minscore}{minimal evidence score for edges to include (see \code{\link[=as_igraph]{as_igraph()}})}

\item{minnodes}{minimal number of nodes in component to include in plot}
}
\value{
a ggplot object
}
\description{
Plot a gene network graph

Plot a gene network graph
}
\examples{
gr <- as_igraph(c("ENSG00000130203", "ENSG00000189058"),  0.9)
plot_graph(gr, c("ENSG00000130203"))

plot_pathway(
    c("ENSG00000130203", "ENSG00000189058"),
    c("ENSG00000130203", "ENSG00000189058"),
    0.9
)

}
